factor out notes table and share between notes and reallocation
